+++ src/main/java/com/dev/cinema/controllers/MovieSessionController.java
@@ -1,6 +1,6 @@
 package com.dev.cinema.controllers;
 
-import com.dev.cinema.controllers.mappers.MovieSessionMapper;
+import com.dev.cinema.service.mappers.MovieSessionMapper;
 import com.dev.cinema.model.MovieSession;
 import com.dev.cinema.model.dto.MovieSessionRequestDto;
 import com.dev.cinema.model.dto.MovieSessionResponseDto;
@@ -31,14 +31,14 @@ public class MovieSessionController {
     public MovieSession addMovieSession(
             @RequestBody MovieSessionRequestDto movieSessionRequestDto) {
         return movieSessionService.add(
-                movieSessionMapper.fromRequestToSession(movieSessionRequestDto));
+                movieSessionMapper.mapToMovie(movieSessionRequestDto));
     }
 
     @GetMapping("/available")
     public List<MovieSessionResponseDto> getAvailableMovieSessions(
             @RequestParam Long movieId, @RequestParam LocalDate date) {
         return movieSessionService.findAvailableSessions(movieId, date).stream()
-                .map(movieSessionMapper::toResponseFromSession)
+                .map(movieSessionMapper::mapToDto)
                 .collect(Collectors.toList());
     }
 }
