+++ src/main/java/com/dev/cinema/service/mappers/TicketMapper.java
@@ -1,4 +1,4 @@
-package com.dev.cinema.controllers.mappers;
+package com.dev.cinema.service.mappers;
 
 import com.dev.cinema.model.Ticket;
 import com.dev.cinema.model.dto.TicketResponseDto;
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class TicketMapper {
 
-    public TicketResponseDto fromEntityToDto(Ticket ticket) {
+    public TicketResponseDto mapToDto(Ticket ticket) {
         TicketResponseDto ticketResponseDto = new TicketResponseDto();
         ticketResponseDto.setTicketId(ticket.getId());
         ticketResponseDto.setUserId(ticket.getUser().getId());
